v3.5.1 unix
git-svn-id: svn://svn.code.sf.net/p/steemsse/code/trunk/steemsse@29 dd3fbd3b-ef5d-4b78-98e8-da144fa02e5b

+++ steem/code/SSE/SSE.cpp
@@ -39,7 +39,8 @@ FARPROC WINAPI MyLoadFailureHook(dliNotification dliNotify,DelayLoadInfo * pdli)
 TOsdControl OsdControl;
 
 extern bool osd_show_scrollers,osd_shown_scroller;
-extern DWORD timer,osd_scroller_finish_time;
+//extern DWORD timer,osd_scroller_finish_time;
+extern long timer,osd_scroller_finish_time;
 extern long col_yellow[2],col_blue,col_red,col_green,col_white;
